Moved business logic to its respected service layer
