+++ src/main/java/com/calzyr/services/user/UserService.java
@@ -1,4 +1,4 @@
-package com.calzyr.services;
+package com.calzyr.services.user;
 
 import com.calzyr.dtos.event.EventResponseDTO;
 import com.calzyr.dtos.user.UserResponseDTO;
@@ -7,12 +7,9 @@ import com.calzyr.entities.user.User;
 import com.calzyr.repositories.EventRepository;
 import com.calzyr.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
 
 import java.sql.Timestamp;
 import java.time.Instant;
